+++ pyxel/exposure/exposure.py
@@ -9,8 +9,8 @@
 """Observation class and functions."""
 
 import logging
-import operator
-from collections.abc import Mapping, Sequence
+from collections import defaultdict
+from collections.abc import Sequence
 from typing import TYPE_CHECKING, Optional, Union
 
 import numpy as np
@@ -19,7 +19,7 @@ from datatree import DataTree
 from tqdm.auto import tqdm
 
 from pyxel import __version__
-from pyxel.data_structure import Charge, Image, Photon, Pixel, Signal
+from pyxel.data_structure import Charge, Image, Photon, Photon3D, Pixel, Scene, Signal
 from pyxel.pipelines import Processor, ResultId, get_result_id, result_keys
 from pyxel.util import set_random_seed
 
@@ -68,6 +68,7 @@ class Exposure:
         """TBW."""
         self._pipeline_seed = value
 
+    # TODO: This function will be deprecated
     def run_exposure(self, processor: Processor) -> "xr.Dataset":
         """Run an observation pipeline.
 
@@ -192,8 +193,7 @@ def run_exposure_pipeline(
 
         keys = result_keys(result_type)
 
-        unstacked_result: Mapping[str, list] = {key: [] for key in keys}
-
+        unstacked_result: dict[str, list] = defaultdict(list)
         i: int
         time: float
         step: float
@@ -219,19 +219,33 @@ def run_exposure_pipeline(
                 outputs.save_to_file(processor)
 
             for key in keys:
-                if key == "data":
+                if key in ("data", "scene"):
                     continue
 
-                unstacked_result[key].append(
-                    np.array(operator.attrgetter(key)(detector))
-                )
+                obj: Union[
+                    Scene, Photon, Photon3D, Pixel, Image, Signal, Charge
+                ] = getattr(detector, key)
+
+                # TODO: Is this necessary ?
+                if not isinstance(
+                    obj, (Photon, Photon3D, Pixel, Image, Signal, Charge)
+                ):
+                    raise TypeError(
+                        f"Wrong type from attribute 'detector.{key}'. Type: {type(obj)!r}"
+                    )
+
+                if obj._array is not None:
+                    data_arr: np.ndarray = np.array(obj)
+                    unstacked_result[key].append(data_arr)
 
             if progressbar:
                 pbar.update(1)
 
         # TODO: Refactor '.result'. See #524
         processor.result = {
-            key: np.stack(unstacked_result[key]) for key in keys if key != "data"
+            key: np.stack(value)
+            for key, value in unstacked_result.items()
+            if key != "data"
         }
 
         if progressbar:
@@ -261,16 +275,18 @@ def _extract_datatree(detector: "Detector", keys: Sequence[ResultId]) -> DataTre
     --------
     >>> _extract_datatree(
     ...     detector=detector,
-    ...     keys=["photon", "charge", "pixel", "signal", "image", "data"],
+    ...     keys=["photon", "photon3d", "charge", "pixel", "signal", "image", "data"],
     ... )
     DataTree('None', parent=None)
-        Dimensions:  (time: 1, y: 100, x: 100)
+        Dimensions:  (time: 1, y: 100, x: 100, wavelength: 201)
         Coordinates:
           * time     (time) float64 1.0
           * y        (y) int64 0 1 2 3 4 5 6 7 8 9 10 ... 90 91 92 93 94 95 96 97 98 99
           * x        (x) int64 0 1 2 3 4 5 6 7 8 9 10 ... 90 91 92 93 94 95 96 97 98 99
+          * wavelength  (wavelength) float64 500.0 502.0 504.0 ... 896.0 898.0 900.0
         Data variables:
             photon   (time, y, x) float64 1.515e+04 1.592e+04 ... 1.621e+04 1.621e+04
+            photon3d    (time, wavelength, y, x) float64 0.0 0.0 0.0 0.0 ... 0.0 0.0 0.0
             charge   (time, y, x) float64 1.515e+04 1.592e+04 ... 1.621e+04 1.621e+04
             pixel    (time, y, x) float64 1.515e+04 1.592e+04 ... 1.621e+04 1.621e+04
             signal   (time, y, x) float64 0.04545 0.04776 0.04634 ... 0.04862 0.04862
@@ -290,9 +306,12 @@ def _extract_datatree(detector: "Detector", keys: Sequence[ResultId]) -> DataTre
         if key.startswith("data") or key.startswith("scene"):
             continue
 
-        obj: Union[Photon, Pixel, Image, Signal, Charge] = getattr(detector, key)
+        obj: Union[Photon, Photon3D, Pixel, Image, Signal, Charge] = getattr(
+            detector, key
+        )
 
-        if not isinstance(obj, (Photon, Pixel, Image, Signal, Charge)):
+        # TODO: Is this necessary ?
+        if not isinstance(obj, (Photon, Photon3D, Pixel, Image, Signal, Charge)):
             raise TypeError(
                 f"Wrong type from attribute 'detector.{key}'. Type: {type(obj)!r}"
             )